requirements.txt
Flask==2.0.2
supabase-py==0.2.6
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
